+++ src/renderer/patches/index.ts
@@ -8,3 +8,4 @@
 import "./spellCheck";
 import "./platformClass";
 import "./windowsTitleBar";
+import "./screenShareAudio";
